Add privacy bulletin: Rare accidental logging of network request destination on Psiphon iOS VPN
Also stopped translating the privacy bulletins. The translation load is too high. They can run very long (and should be allowed to be as long as they need to be), they can be very technical (and should be allowed to be as technical as they need to be), and there is an ever-growing number of them (meaning that translation load for new languages will come to be dominated by them).

They can be translated as needed, like we did/do with the old-style blog posts.