Improve accuracy of email handling info
Keeping email addreses out of syslog is extremely difficult, and often necessary for diagnosing issues.

We also don't always send two email responses, and the fact that we're using SES doesn't really affect the theoretical visibility to Amazon -- the server is in AWS, so even SMTP email will be visible.